Factor out _common.py.
